requirements.txt


langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langgraph==0.2.14
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1
gradio==4.42.0
python-dotenv==1.0.1